refactor: simplify snapshot api flow
