+++ src/shared/condense-dom/condense-dom.ts
